Moved level/weapon/consile/saves utils.
