+++ src/engine/core/managers/save/SaveManager.ts
@@ -1,20 +1,20 @@
 import { initializeManager } from "@/engine/core/database";
 import { AchievementsManager } from "@/engine/core/managers/achievements";
 import { ActorInputManager } from "@/engine/core/managers/actor";
-import { AbstractManager } from "@/engine/core/managers/base/AbstractManager";
-import { ReleaseBodyManager } from "@/engine/core/managers/death/ReleaseBodyManager";
+import { AbstractManager } from "@/engine/core/managers/base";
+import { ReleaseBodyManager } from "@/engine/core/managers/death";
 import { EGameEvent, EventsManager } from "@/engine/core/managers/events";
-import { PsyAntennaManager } from "@/engine/core/managers/psy/PsyAntennaManager";
+import { PsyAntennaManager } from "@/engine/core/managers/psy";
 import { IDynamicSaveData } from "@/engine/core/managers/save/save_types";
 import { GameSettingsManager } from "@/engine/core/managers/settings";
-import { SimulationBoardManager } from "@/engine/core/managers/simulation/SimulationBoardManager";
-import { GlobalSoundManager } from "@/engine/core/managers/sounds/GlobalSoundManager";
+import { SimulationBoardManager } from "@/engine/core/managers/simulation";
+import { GlobalSoundManager } from "@/engine/core/managers/sounds";
 import { StatisticsManager } from "@/engine/core/managers/statistics";
-import { SurgeManager } from "@/engine/core/managers/surge/SurgeManager";
+import { SurgeManager } from "@/engine/core/managers/surge";
 import { TaskManager } from "@/engine/core/managers/tasks";
 import { TreasureManager } from "@/engine/core/managers/treasures";
-import { WeatherManager } from "@/engine/core/managers/weather/WeatherManager";
-import { loadDynamicGameSave, saveDynamicGameSave } from "@/engine/core/utils/game";
+import { WeatherManager } from "@/engine/core/managers/weather";
+import { loadDynamicGameSave, saveDynamicGameSave } from "@/engine/core/utils/game_save";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { NetPacket, NetProcessor, Optional, TName } from "@/engine/lib/types";
