+++ src/engine/core/objects/binders/creature/StalkerBinder.ts
@@ -39,13 +39,9 @@ import { assert } from "@/engine/core/utils/assertion";
 import { getObjectCommunity } from "@/engine/core/utils/community";
 import { pickSectionFromCondList, readIniString, TConditionList } from "@/engine/core/utils/ini";
 import { ISmartTerrainJobDescriptor } from "@/engine/core/utils/job";
+import { isUndergroundLevel } from "@/engine/core/utils/level";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import {
-  getObjectStalkerIni,
-  isUndergroundLevel,
-  setupObjectInfoPortions,
-  setupObjectStalkerVisual,
-} from "@/engine/core/utils/object";
+import { getObjectStalkerIni, setupObjectInfoPortions, setupObjectStalkerVisual } from "@/engine/core/utils/object";
 import { ERelation, setClientObjectRelation, setObjectSympathy } from "@/engine/core/utils/relation";
 import {
   emitSchemeEvent,
