+++ src/engine/core/ui/debug/sections/DebugTeleportSection.ts
@@ -5,7 +5,8 @@ import { EGameEvent, EventsManager } from "@/engine/core/managers/events";
 import { SmartTerrain } from "@/engine/core/objects/server/smart_terrain";
 import { AbstractDebugSection } from "@/engine/core/ui/debug/sections/AbstractDebugSection";
 import { DebugTeleportListEntry } from "@/engine/core/ui/debug/sections/DebugTeleportListEntry";
-import { executeConsoleCommand, isGameStarted } from "@/engine/core/utils/game";
+import { executeConsoleCommand } from "@/engine/core/utils/console";
+import { isGameStarted } from "@/engine/core/utils/game";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { isGameVertexFromLevel } from "@/engine/core/utils/position";
 import { getServerObjects } from "@/engine/core/utils/registry";
