+++ src/engine/core/ui/menu/load/LoadDialog.ts
@@ -24,7 +24,7 @@ import {
   getGameSavesList,
   isGameSaveFileExist,
   loadGameSave,
-} from "@/engine/core/utils/game/game_save";
+} from "@/engine/core/utils/game_save";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { resolveXmlFormPath } from "@/engine/core/utils/ui";
 import { gameConfig } from "@/engine/lib/configs/GameConfig";
