+++ src/engine/core/utils/game/index.ts
@@ -1,4 +1,2 @@
-export * from "@/engine/core/utils/game/game_console";
-export * from "@/engine/core/utils/game/game_save";
 export * from "@/engine/core/utils/game/game_check";
 export * from "@/engine/core/utils/game/game_wait";
