+++ src/engine/core/utils/weapon.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "@jest/globals";
 
-import { getObjectActiveWeaponSlot } from "@/engine/core/utils/weapon";
+import { getObjectActiveWeaponSlot, isObjectStrappingWeapon } from "@/engine/core/utils/weapon";
 import { ClientObject } from "@/engine/lib/types";
 import { replaceFunctionMock } from "@/fixtures/jest";
 import { mockClientGameObject } from "@/fixtures/xray";
@@ -21,6 +21,26 @@ describe("object_weapon utils", () => {
 
     replaceFunctionMock(object.active_item, () => mockClientGameObject({ animation_slot: () => 3 }));
     expect(getObjectActiveWeaponSlot(object)).toBe(3);
+  });
+
+  it("isObjectStrappingWeapon should correctly check weapon strap state", () => {
+    const object: ClientObject = mockClientGameObject();
+
+    replaceFunctionMock(object.weapon_strapped, () => true);
+    replaceFunctionMock(object.weapon_unstrapped, () => false);
+    expect(isObjectStrappingWeapon(object)).toBe(false);
+
+    replaceFunctionMock(object.weapon_strapped, () => false);
+    replaceFunctionMock(object.weapon_unstrapped, () => true);
+    expect(isObjectStrappingWeapon(object)).toBe(false);
+
+    replaceFunctionMock(object.weapon_strapped, () => true);
+    replaceFunctionMock(object.weapon_unstrapped, () => true);
+    expect(isObjectStrappingWeapon(object)).toBe(false);
+
+    replaceFunctionMock(object.weapon_strapped, () => false);
+    replaceFunctionMock(object.weapon_unstrapped, () => false);
+    expect(isObjectStrappingWeapon(object)).toBe(true);
   });
 
   it.todo("isObjectWeaponLocked should correctly check weapon locked state");
