+++ src/engine/core/utils/weapon.ts
@@ -6,19 +6,13 @@ import { isWeapon } from "@/engine/core/utils/class_ids";
 import { ClientObject, Optional, TIndex } from "@/engine/lib/types";
 
 /**
- * Get active weapon slot of an object for animating.
+ * Check whether object is strapping weapon.
  *
  * @param object - target client object to check
- * @returns active weapon slot index
+ * @returns whether strapping/unstrapping weapon is in process
  */
-export function getObjectActiveWeaponSlot(object: ClientObject): TIndex {
-  const weapon: Optional<ClientObject> = object.active_item();
-
-  if (weapon === null || object.weapon_strapped()) {
-    return 0;
-  }
-
-  return weapon.animation_slot();
+export function isObjectStrappingWeapon(object: ClientObject): boolean {
+  return !(object.weapon_unstrapped() || object.weapon_strapped());
 }
 
 /**
@@ -54,6 +48,22 @@ export function setObjectBestWeapon(target: ClientObject): void {
   if (bestWeapon && isWeapon(bestWeapon)) {
     target.set_item(object.idle, bestWeapon);
   }
+}
+
+/**
+ * Get active weapon slot of an object for animating.
+ *
+ * @param object - target client object to check
+ * @returns active weapon slot index
+ */
+export function getObjectActiveWeaponSlot(object: ClientObject): TIndex {
+  const weapon: Optional<ClientObject> = object.active_item();
+
+  if (weapon === null || object.weapon_strapped()) {
+    return 0;
+  }
+
+  return weapon.animation_slot();
 }
 
 /**
