+++ src/engine/scripts/declarations/effects/game.ts
@@ -3,8 +3,8 @@ import { game, get_hud, StaticDrawableWrapper } from "xray16";
 import { getPortableStoreValue, setPortableStoreValue } from "@/engine/core/database";
 import { ItemUpgradesManager } from "@/engine/core/managers/interface/ItemUpgradesManager";
 import { extern } from "@/engine/core/utils/binding";
-import { executeConsoleCommand } from "@/engine/core/utils/game/game_console";
-import { createGameAutoSave } from "@/engine/core/utils/game/game_save";
+import { executeConsoleCommand } from "@/engine/core/utils/console";
+import { createGameAutoSave } from "@/engine/core/utils/game_save";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { consoleCommands } from "@/engine/lib/constants/console_commands";
 import { ACTOR_ID } from "@/engine/lib/constants/ids";
